pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"


[project]
name = "symphony"
authors = [
    {name = "Ameya Daigavane", email = "ameya.d.98@gmail.com"},
    {name = "Song Kim", email = "songk@mit.edu"},
    {name = "Mario Geiger", email = "geiger.mario@gmail.com"},
]
description = "Symphony"
readme = "README.md"
requires-python = ">=3.10"
keywords = ["generative", "equivariant", "molecules"]
license = {text = "BSD-3-Clause"}
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "e3nn_jax",
    "numpy",
    "jax",
    "flax",
    "dm-haiku",
    "jraph",
    "clu==0.0.9",
    "absl-py",
    "chex",
    "optax",
    "scipy",
    "tensorflow==2.13.0",
    "rdkit",
    "wandb",
    "py3dmol",
    "posebusters",
    "sh",
    "distrax",
    "tensorflow-probability==0.21.0",
    "mace-jax @ git+https://github.com/ACEsuit/mace-jax",
    "nequip-jax @ git+https://github.com/mariogeiger/nequip-jax",
    "allegro-jax @ git+https://github.com/mariogeiger/allegro-jax",
]
dynamic = ["version"]

[project.optional-dependencies]
plotting = ["pandas", "matplotlib", "plotly"]

[tool.setuptools.packages]
find = {}

[tool.black]
target-version = ['py39']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | .nox
)/
'''
